[mandatory] mini_env working
