+++ sources/builtin/ft_env.c
@@ -6,7 +6,7 @@
 /*   By: feralves <feralves@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/30 17:37:53 by feralves          #+#    #+#             */
-/*   Updated: 2023/04/10 19:14:46 by feralves         ###   ########.fr       */
+/*   Updated: 2023/04/10 19:19:51 by feralves         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,5 +30,6 @@ void	ft_env(t_token *token, t_mini_env *mini_env)
 			ft_printf ("%s=%s\n", aux->name, aux->value);
 			aux = aux->next;
 		}
+		ft_printf ("%s=%s\n", aux->name, aux->value);
 	}
 }
